[MIG] account_global_discount: Migration to 19.0
