+++ account_global_discount/models/__init__.py
@@ -1,2 +1,4 @@
+from . import account_invoice_global_discount
+from . import account_move_line
 from . import account_move
 from . import global_discount
