+++ account_global_discount/views/global_discount_views.xml
@@ -43,6 +43,6 @@
         action="base_global_discount.action_global_discount_tree"
         name="Global Discounts"
         groups="base_global_discount.group_global_discount"
-        parent="account.account_management_menu"
+        parent="account.account_invoicing_menu"
     />
 </odoo>
